refactor: split workspace report computation
